Bugfix - size attribute cannot be applied to EXFAT_CHAINED_EXTENTS?
It seems that the size attribute could be used, because the resulting
location of the next structure after parsing this item is always at
a fixed offset of four bytes.  This is because those four bytes
are just a first pointer, and the chain is followed, then parsing
should continue at +4 bytes.  This matches the documentation.

This is ***NOT*** how it actually works.
Apparently, using the `size=` attribute also prevents local variables
from being able to access struct parameters?  Need to investigate this
oddity more, and perhaps open another bug with Sweetscape support....