Reduce local variables.
This is a bit of a hack.
Structs are allowed to have parameters.
It's a pain to ensure all local variables have <hidden=true>.
But, if the struct simply calls a function that does the same
stuff, when the function returns, the local variables are all
de-allocated.  AWESOME!  This makes complex struct writing
slightly more sane.